MapWindow/GlueMapWindowDisplayMode: Fix OffsetHistory average calculation regression
Regression introduced in commit f989ed4caf5beef8c904229e47894bde053c43de

+++ src/MapWindow/GlueMapWindowDisplayMode.cpp
@@ -56,8 +56,8 @@ OffsetHistory::average() const
   }
 
   RasterPoint avg;
-  avg.x = x / offsets.size();
-  avg.y = y / offsets.size();
+  avg.x = x / (int) offsets.size();
+  avg.y = y / (int) offsets.size();
 
   return avg;
 }
